Fix golangci-lint findings in v5 port
- r9nano: extract applyHBMTimings helper from hbmDRAMSpec (funlen: 44>40)
- addresstranslator: extract processTranslationRsp from parseTranslation
  (funlen: 63>60)
- driver/api_test: drop unnecessary uint64() casts on mem.MB (unconvert)

+++ amd/driver/api_test.go
@@ -85,7 +85,7 @@ var _ = ginkgo.Describe("Driver async API execution", func() {
 		ptr := driver.AllocateMemory(context, 1*mem.MB)
 
 		Expect(context.buffers).To(HaveLen(1))
-		Expect(context.buffers[0].size).To(Equal(uint64(1 * mem.MB)))
+		Expect(context.buffers[0].size).To(Equal(1 * mem.MB))
 		Expect(context.buffers[0].vAddr).To(Equal(ptr))
 		Expect(context.buffers[0].freed).To(BeFalse())
 		Expect(context.buffers[0].l2Dirty).To(BeFalse())
@@ -97,7 +97,7 @@ var _ = ginkgo.Describe("Driver async API execution", func() {
 		ptr := driver.AllocateUnifiedMemory(context, 1*mem.MB)
 
 		Expect(context.buffers).To(HaveLen(1))
-		Expect(context.buffers[0].size).To(Equal(uint64(1 * mem.MB)))
+		Expect(context.buffers[0].size).To(Equal(1 * mem.MB))
 		Expect(context.buffers[0].vAddr).To(Equal(ptr))
 		Expect(context.buffers[0].freed).To(BeFalse())
 		Expect(context.buffers[0].l2Dirty).To(BeFalse())
